Revert "[Change] 장미 sprite 프리팹 수정 및 카드 아트 스크립트 추가"
This reverts commit 9911c503fe9fec169e4761629b3579714ab572dc.
