+++ Assets/Prefabs/Cards/02. AdjustCard/FloatCard.prefab
@@ -147,11 +147,6 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: fe4a27cc58c5b4998b897f6262cfeda1, type: 3}
---- !u!4 &1007093746756822017 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 1263317207295340012, guid: fe4a27cc58c5b4998b897f6262cfeda1, type: 3}
-  m_PrefabInstance: {fileID: 2049663550649993709}
-  m_PrefabAsset: {fileID: 0}
 --- !u!4 &5602665765175406042 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 5886614794521977911, guid: fe4a27cc58c5b4998b897f6262cfeda1, type: 3}
@@ -167,78 +162,3 @@ GameObject:
   m_CorrespondingSourceObject: {fileID: 7277805989260689589, guid: fe4a27cc58c5b4998b897f6262cfeda1, type: 3}
   m_PrefabInstance: {fileID: 2049663550649993709}
   m_PrefabAsset: {fileID: 0}
---- !u!1001 &5165374655450719016
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 1007093746756822017}
-    m_Modifications:
-    - target: {fileID: 1988271296621881820, guid: c172fbe93e4863f46a661cb376658be7, type: 3}
-      propertyPath: m_RootOrder
-      value: 5
-      objectReference: {fileID: 0}
-    - target: {fileID: 1988271296621881820, guid: c172fbe93e4863f46a661cb376658be7, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1988271296621881820, guid: c172fbe93e4863f46a661cb376658be7, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1988271296621881820, guid: c172fbe93e4863f46a661cb376658be7, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: -0.1
-      objectReference: {fileID: 0}
-    - target: {fileID: 1988271296621881820, guid: c172fbe93e4863f46a661cb376658be7, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 1988271296621881820, guid: c172fbe93e4863f46a661cb376658be7, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1988271296621881820, guid: c172fbe93e4863f46a661cb376658be7, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1988271296621881820, guid: c172fbe93e4863f46a661cb376658be7, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1988271296621881820, guid: c172fbe93e4863f46a661cb376658be7, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1988271296621881820, guid: c172fbe93e4863f46a661cb376658be7, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1988271296621881820, guid: c172fbe93e4863f46a661cb376658be7, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6460080778222183648, guid: c172fbe93e4863f46a661cb376658be7, type: 3}
-      propertyPath: m_Name
-      value: RoseSprite
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: c172fbe93e4863f46a661cb376658be7, type: 3}
---- !u!1 &2164483043122808776 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 6460080778222183648, guid: c172fbe93e4863f46a661cb376658be7, type: 3}
-  m_PrefabInstance: {fileID: 5165374655450719016}
-  m_PrefabAsset: {fileID: 0}
---- !u!114 &7414206931586211842
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2164483043122808776}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 329b682c837acef458a250471cb30ac0, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  highLight: {fileID: 8686928968484850008}
