+++ Assets/Prefabs/Cards/02. AdjustCard/MoveRose.cs.meta
