--- Assets/Prefabs/UI/RoseSprite.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c172fbe93e4863f46a661cb376658be7
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
