--- Assets/Scripts/Card/CardArtEffect.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 17581d054aa6da44d810ab5653cc9f39
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
