--- Assets/Scripts/Card/CardArtEffect/BouncyRose.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d18b29007cbbc1844a283c8cbb5f4fa5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
