--- Assets/Scripts/Card/CardArtEffect/FloatRose.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 329b682c837acef458a250471cb30ac0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
