--- Assets/Scripts/Card/CardArtEffect/LongRose.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d76147dc932742443bba53219adbcf5b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
